Use target size limit in strncpy() in adapter.c.
This fixes the following compiler warnings and potentially avoid a buffer overflow:

adapter.c: In function ‘getAdapterConfig3’:
adapter.c:510:63: warning: argument to ‘sizeof’ in ‘strncpy’ call is the same expression as the source; did you mean to use the size of the destination? [-Wsizeof-pointer-memaccess]
  510 |     strncpy (a->product, (char *) pinfo->product_name, sizeof (pinfo->product_name));
      |                                                               ^
adapter.c: In function ‘getAdapterConfig5’:
adapter.c:668:63: warning: argument to ‘sizeof’ in ‘strncpy’ call is the same expression as the source; did you mean to use the size of the destination? [-Wsizeof-pointer-memaccess]
  668 |     strncpy (a->product, (char *) pinfo->product_name, sizeof (pinfo->product_name));
      |                                                               ^